Modularize app implementation
